Adding in priest functionality. Vampire abilities have no known bugs.

+++ src/me/matterz/supernaturals/commands/SNCommandPower.java
@@ -35,6 +35,6 @@ public class SNCommandPower extends SNCommand {
 		SupernaturalsPlugin.instance.getSuperManager();
 		SuperNPlayer snplayer = SupernaturalManager.get(senderPlayer);
 				
-		this.sendMessage("Your current power level is: " + snplayer.getPower());
+		this.sendMessage("Your current power level is: " + (int) snplayer.getPower());
 	}
 }
